Small refactoring around prompt objects
Delegate class `.call` to an instance of the prompt object. Cleans up the interface a little bit and
pushes the one prompt instance into the base class.

Also - allow quitting the program at any point with the `esc` key.
AND - allow using j/k for up and down in prompts as well.