Initial credential saving too convoluted - pull back
Currently it's not in a place where I felt like it worked. Credential saving, for now,
is a decision that if anyone wants to use it, they can save the target file somewhere
safe in their filesystem. `chmod 600 ~/location/of/file` should be good enough right
now to keep it safe. The main connection string _may_ already exist in the toml config
file -- if one would like -- so it's not like we're already operating in the most
secure setup.

In the future there's probably a place where the creds file is encrypted in its
entirety before it's accessed but the solution I had until now was just not good enough.