Store connection strings for new db's and users
... this allows for quick retrieval so that I may connect to those db's w/created users.

Why?
====

Previously, when self-hosting applications that require its own
PG database, I'll create one, create a user, set it up, connect everything up
and sometimes will forget to put it in my password manager.

Most of the utility of this tool is that I want something VERY small and out
of the way that will do the common tasks that shouldn't require something like
PGAdmin. Do a couple of things. Don't make me spin up yet another container
to do something I might use every few weeks (if that).

So, for me, if I keep a file in my $HOME with a bunch of connection strings,
chmod'ed to 600, then that's enough.

The initial plan was to encrypt everything in there but, for now that may be
overkill.

What:
======

Create store and retrieve connection string command objects.
Call out to those classes when creating a db or assigning a user privileges to them.
Update config file.