Use prioritized locations for primary PG connection
Why?
====

You can set your main (admin[istrative]) PG connection string in the generated
config file. You can also set it via an environment variable (`$DATABASE_URL`).
The order of prioritization should be 1) the ENV and 2) the config file.

When trying out the gem on a fresh machine there is also a bug. It fails if you
do not have a DATABASE_URL env set and you run that `--init` flag.

This fixes that as well.

What?
=====

Remove the constant in `Pgchief` that is set via the ENV.
Delegate all work to access that connection string through the `Config` class.