Backup database (local and remote)
Why?
====

Backing things up locally - no problem.
Backing things up to a remote store? Even better!

What?
=====

Add some options in the config for s3. If everything necessary is provided,
throw all of that stuff into an S3-specific object that can be used in the
operations that would utilize it.