Allow the user to add pg extensions to the specific database
When setting up a rails app that had a couple of PG extensions in the schema.rb file
I ran into the issue of not having the correct permissions to do so. This change
will now allow that created user to do that, without owning the DB or having SUPERUSER
privileges.